requirements.txt
django==1.8
djangorestframework==3.0.5
mongoengine==0.9.0
pymongo==2.8
django-rest-framework-mongoengine==2.0.2
